CLI

Documentation

menu :: IO ()

Displays the main menu and processes user input.

This function shows the available options to the user and processes the chosen option accordingly.

processOption :: String -> IO ()

Processes the selected option from the main menu.

Parameters: - String : The option selected by the user.

getCSVFilePath :: String -> FilePath

Returns the file path associated with the model name.

Parameters: - String : The model's name.

Returns: - FilePath : The corresponding CSV file path.

classificationSubmenu :: Map String Double -> Map String Double -> IO ()

Submenu for classifying individual messages.

This function displays a submenu where users can choose to classify a message or return to the main menu.

Parameters: - `Map.Map String Double` : Word probabilities for ham messages. - `Map.Map String Double` : Word probabilities for spam messages.

loop :: Map String Double -> Map String Double -> IO ()

Loop for entering messages to be classified.

Parameters: - `Map.Map String Double` : Word probabilities for ham messages. - `Map.Map String Double` : Word probabilities for spam messages.

trainingManualLoop :: FilePath -> IO ()

Loops for entering spam and ham messages and saving them for manual training.

Parameters: - FilePath : The file path where the data will be saved.

trainingManualSubmenu :: FilePath -> String -> IO ()

Submenu for training manually.

Parameters: - FilePath : The file path of the CSV file to save the training data. - String : The model's name.

askPath :: IO String

Asks for the path of the model to be added.

Returns: - `IO String` : The file path, or "unknown" if exit was entered.

addNewModelSubmenu :: IO ()

Submenu for adding a new model.

This submenu allows the user to provide a model name and file path to add a new model.

reusingPreviousModelSubmenu :: IO ()

Submenu for reusing previously saved models.

This function lists the available models and allows the user to select one to reuse.

removeModelSubmenu :: IO ()

Removes an existing model, except for default models.

This function allows the user to remove models from the system, but default models cannot be deleted.