Intro

Documentation

typeWriter :: String -> IO ()

Simulates a typing effect by printing each character with a small delay. The typing stops if the user presses any key, otherwise it continues to display the string.

Parameters: - msg : The string to be displayed with the typing effect.

Return: - `IO ()` : Action that produces the typing effect.

animatedLogo :: IO ()

Displays the system's logo with an animated typing effect. Each line of the logo is displayed one by one, with a delay in between. The user can interrupt the animation by pressing any key, which will clear the terminal.

Return: - `IO ()` : Action that displays the animated logo.

waitForAnyKey :: IO ()

Waits for the user to press any key to continue. The input is not buffered, and the program will proceed as soon as any key is pressed.

Return: - `IO ()` : Action that waits for the user to press a key.

showIntro :: IO ()

Displays the introductory animation with the system's logo and a welcome message. After the animation, it shows a brief description of the program's functionality and waits for the user to press any key before proceeding.

Return: - `IO ()` : Action that shows the introduction and waits for user input.

showOut :: IO ()

Displays a final message with a typing effect when the user exits the program. The message expresses gratitude and bids farewell to the user.

Return: - `IO ()` : Action that shows the final message before exiting.