Metric

Documentation

showAccuracy :: FilePath -> IO Double

Calculate the model accuracy. Parameters: - FilePath: path of file. Return: - 'IO Double': value of accuracy.

accuracyRecursion :: [(String, FilePath)] -> IO ()

Performs recursion to show the accuracy of each model. Parameters: - '[(String, FilePath)]': A vector of map with the key (model name) and value (path of file). Return: - 'IO()': a little explanation of the algorithm calculation and a table with name, accuracy and classification of each model.

accuracyCSVs :: FilePath -> IO ()

Downloads the list of JSON templates to calculate precision in precisionRecursion. Parameters: - FilePath: Path with the paths and names of the data files. Return: - 'IO()': the return of accuracyRecursion.

modelClassification :: Double -> String

Ranges with each accuracy rating. Parameters: - Double: Value of accuracy. Return: - String: the classification according to accuracy.