Model

Documentation

countWords :: Vector MyRecord -> Map String Int -> Map String Int -> (Int, Int) -> (Map String Int, Map String Int, Int, Int)

Counts the words for each class (ham or spam) based on the message records.

This function iterates over a vector of records, counting the occurrences of each word for both ham and spam messages. It returns updated maps of word counts for both classes and the total number of ham and spam messages.

Parameters: - records : A vector (`V.Vector MyRecord`) containing the message records. - hamWords : A map (`Map.Map String Int`) containing the word counts for ham messages. - spamWords : A map (`Map.Map String Int`) containing the word counts for spam messages. - `(hamCount, spamCount)` : A tuple containing the number of ham and spam messages.

Return: - A tuple containing: - Updated hamWords and spamWords maps. - Updated hamCount and spamCount values.

calculateWordProbabilities :: Map String Int -> Int -> Map String Int -> Int -> Map String Double -> Map String Double

Calculates the probabilities of words for each class (ham and spam).

This function calculates the probability of each word occurring in the given class using Laplace smoothing. It then computes the ratio of the probability of the word in the ham class over its probability in the spam class.

Parameters: - wordsCount : A map (`Map.Map String Int`) containing word counts for a given class. - totalCount : The total number of words in the given class. - otherWordsCount : A map (`Map.Map String Int`) containing word counts for the other class. - otherCount : The total number of words in the other class. - wordProbs : A map (`Map.Map String Double`) containing previously calculated word probabilities.

Return: - A map (`Map.Map String Double`) containing the updated word probabilities.