Utils

Documentation

data MyRecord

Represents a record from a CSV file, containing a label and a message.

Fields: - label : A string representing the type of message (e.g., ham or spam). - message : The actual message content.

Constructors

MyRecord

Fields

label :: String

Type of message (ham or spam)

message :: String

Message content

type ModelMap = Map String FilePath

Type alias for the model map, which stores models with their associated file paths.

ModelMap is a map that associates model names (strings) with their file paths.

readCSV :: FilePath -> IO (Either String (Vector MyRecord))

Reads a CSV file and decodes its contents into a vector of MyRecord.

Parameters: - filePath : The path to the CSV file to be read.

Return: - `Either String (V.Vector MyRecord)` : Either an error message if the file can't be decoded, or the vector of records.

divideDataset :: Vector MyRecord -> (Vector MyRecord, Vector MyRecord)

Divides a dataset into training and test sets (70% training, 30% test).

Parameters: - records : A vector of MyRecord containing all the data.

Return: - `(V.Vector MyRecord, V.Vector MyRecord)` : A tuple containing the training and test datasets.

divideCsvTrainingTest :: FilePath -> Vector MyRecord -> IO (Vector MyRecord, Vector MyRecord)

Divides the dataset into training and test sets based on a predefined CSV file or uses a default dataset.

Parameters: - filePath : The path to the CSV file. - records : A vector of MyRecord containing the data to be divided.

Return: - `(V.Vector MyRecord, V.Vector MyRecord)` : The training and test datasets.

downloadDefault :: IO (Vector MyRecord)

Downloads the default dataset and returns it as a vector of MyRecord.

Return: - `V.Vector MyRecord` : The default dataset.

clearTerminal :: IO ()

Clears the terminal screen.

This function clears the terminal based on the operating system.

flushOutput :: IO ()

Flushes the standard output buffer.

This function ensures that all pending output is written to the terminal.

saveToCSV :: FilePath -> String -> String -> IO ()

Saves a message along with its classification to a CSV file.

Parameters: - fileName : The file to save the data. - classification : The classification of the message (e.g., ham or spam). - message : The message to be saved.

This function appends the classification and message to the CSV file and clears the terminal.

loadModelMap :: FilePath -> IO ModelMap

Loads a JSON file containing the models and returns them as a ModelMap.

Parameters: - path : The path to the JSON file containing the models.

Return: - ModelMap : A map of model names to file paths.

printModels :: ModelMap -> IO ()

Show the availabels models. Parameters: - ModelMap: Map with the models names. Return: - 'IO ()': Show availabels models

saveModelToJSON :: String -> FilePath -> IO ()

Update and save the models list. Parameters: - String: Models name. - FilePath: path of model. Return: - 'IO ()': message showing that model has been saved.

ensureCSVExtension :: String -> String

Ensure that the csv file has the suffix .csv. Parameters: - String: file name. Return: - String: the file name with .csv